Tidy up and improve docs for runnable queries
