+++ azuki-core/src/main/kotlin/com/anaplan/engineering/azuki/core/system/Query.kt
@@ -4,77 +4,9 @@ interface QueryFactory
 
 object NoQueryFactory : QueryFactory
 
-
 interface Query<T> : ReifiedBehavior
 
 interface DerivedQuery<T>
-
-/**
- * A query that can be run on an environment.
- */
-interface RunnableQuery<E, CF : CheckFactory, T> : Query<T> {
-
-    /**
-     * Run a query on an environment to return an answer.
-     */
-    fun run(environment: E): Answer<T, CF>
-}
-
-/**
- * Narrows a Query to a RunnableQuery.
- */
-fun <E, CF : CheckFactory, T> Query<T>.ensureRunnable() = let {
-    require(this is RunnableQuery<*, *, T>) {
-        if (this is UnsupportedQuery) "Query is unsupported" else "Query class is not runnable: ${this::class.simpleName}"
-    }
-    @Suppress("UNCHECKED_CAST") (this as RunnableQuery<E, CF, T>)
-}
-
-
-/**
- * A derived query that can be run on an environment.
- */
-fun interface RunnableDerivedQuery<E, CF : CheckFactory, T> : DerivedQuery<T> {
-
-    /**
-     * Derives a list of queries from this derived query based on the state of the environment.
-     */
-    fun derive(environment: E): List<RunnableQuery<E, CF, *>>
-}
-
-/**
- * Adapts a list of runnable queries into a derived query.
- *
- * The main use of this adapter is to form the final layer of a nested quantification.
- */
-data class ListRunnableDerivedQuery<E, CF : CheckFactory, T>(val queries: List<RunnableQuery<E, CF, *>>) :
-    RunnableDerivedQuery<E, CF, T> {
-
-    override fun derive(environment: E) = queries
-
-    companion object {
-
-        fun <E, CF : CheckFactory, T> fromQueries(queries: List<Query<*>>) =
-            ListRunnableDerivedQuery<E, CF, T>(queries.map { it.ensureRunnable() })
-    }
-}
-
-/**
- * Implements a runnable for-all query.
- *
- * The answers from the driver query will be used to produce another layer of derivation, which will then be derived,
- * and so on until we reach a base case (such as a ListRunnableDerivedQuery).
- */
-class ForallRunnableDerivedQuery<E, CF : CheckFactory, T, C : Collection<T>>(
-    val driver: RunnableQuery<E, CF, C>, val derivedQueryFactory: (T) -> RunnableDerivedQuery<E, CF, *>
-) : RunnableDerivedQuery<E, CF, T> {
-
-    override fun derive(environment: E): List<RunnableQuery<E, CF, *>> {
-        val answers = driver.run(environment).value
-        return answers.flatMap { t -> derivedQueryFactory(t).derive(environment) }
-    }
-}
-
 
 class UnsupportedQuery<T> : Query<T> {
